The 2nd incarnation of dhmk.
Read qt-kde-team/2/README for full description of its features. Some features
are not implemented yet (marked as TODO in the README). What's more, all
current snippets in qt-kde-team/1/ need porting to dhmk (as make snippets or dh
addons).

This version of dhmk is not well tested but proof of concept should be there.
Unlike previous version of dhmk, this version no longer supports
{target}_override_{dh_command} overrides.