Relay environment changes made by dh addons to debian/rules.
Add code which detects if dh addons made any changes to the environment and
relays them (if any) to the environment of debian/rules (via dhmk_rules.mk).
This is similar to dh(1) which while being perl based inherits dh addon
environment changes automagically.

This allows pkgkde-symbolshelper dh addon to work with dhmk.