Add initial version of the DLRestrictions library.
This commit implements DLRestrictions library (dlrestrictions/ subdir) and adds
new binary packages for it (libdlrestrictions1 and libdlrestrictions-dev
respectively).

DLRestrictions library that can be used to check compatibility of the external
shared library (or plugin) against other libraries currently loaded in the
global symbol object. It might help to avoid symbol clashes and provide means
for graceful failure detection in advance before the first dlsym() call.
Otherwise, usage of incompatible library might result in crashes or other
unpredictable failures of the application.

A shared library that is protected by DLRestrictions provides a special symbol
which defines conditions under which the library may be safely used. At the
moment, the only restriction that may be imposed controls whether multiple
SONAMEs of the same library can safely co-exist in the same symbol space.
Shared library dependencies are also checked for compatibility.